+++ Code/CrowdControlEffect.cs
@@ -214,7 +214,7 @@ public abstract class CrowdControlEffect : IFeature
         if (_isTimedEffect && !requested)
         {
             // Let Crowd Control know that the timed effect has finished
-            CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, Id, CrowdControlResponseStatus.Finished);
+            CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, this, CrowdControlResponseStatus.Finished);
         }
 
         SendStopMessage();
@@ -243,7 +243,7 @@ public abstract class CrowdControlEffect : IFeature
         IsPaused = true;
 
         // Let Crowd Control know that the effect has been paused
-        CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, Id, CrowdControlResponseStatus.Paused);
+        CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, this, CrowdControlResponseStatus.Paused);
     }
 
     public void Resume()
@@ -257,7 +257,7 @@ public abstract class CrowdControlEffect : IFeature
         IsPaused = false;
 
         // Let Crowd Control know that the effect has been unpaused
-        CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, Id, CrowdControlResponseStatus.Resumed);
+        CrowdControlMod.GetInstance().QueueResponseToCrowdControl(_netId, this, CrowdControlResponseStatus.Resumed);
     }
 
     /// <summary>